requirements.txt
flask-restful
flask
flask-mongoengine
flask-jwt-extended
deepface
matplotlib
